requirements.txt
bcrypt==3.2.2
cryptography
fastapi
PyMySQL
python-jose
python-multipart
SQLAlchemy
pyodbc
uvicorn
passlib
pytest
httpx
pytest-asyncio
aiofiles
jinja2
authlib
httpx
python-dotenv
alembic
pillow
pydantic[email]